requirements.txt
discord.py
mysql-connector-python
python-dotenv
ics
requests
pytz